+++ lib/puppet/type/windows_firewall_ipsec_rule.rb
@@ -154,14 +154,22 @@ Puppet::Type.newtype(:windows_firewall_ipsec_rule) do
     desc "Gets the main mode rules that are associated with the given phase 1 authentication set to be created"
     newvalues(:none, :computerkerberos, :anonymous)
 
-    defaultto :none
+    defaultto do
+      unless @resource[:inbound_security] == 'require' or @resource[:inbound_security] == 'request' or @resource[:outbound_security] == 'require' or @resource[:outbound_security] == 'request'
+        :none
+      end
+    end
   end
 
   newproperty(:phase2auth_set) do
     desc "Gets the IPsec rules that are associated with the given phase 2 authentication set to be created"
     newvalues(:none, :userkerberos)
 
-    defaultto :none
+    defaultto do
+      unless @resource[:inbound_security] == 'require' or @resource[:inbound_security] == 'request' or @resource[:outbound_security] == 'require' or @resource[:outbound_security] == 'request'
+        :none
+      end
+    end
   end
 
   newparam(:name) do